fix: resolve clippy warnings and compilation errors
- Fix sort_by closure types in evo optimizer (&Box<dyn Genome>)
- Fix range type mismatch in PrimeGenerator sieve
- Remove unnecessary let_and_return in za function
- Change ptr_arg to slice in create_segments_from_geo
- All 32 tests pass

+++ src/evo/mod.rs
@@ -92,8 +92,8 @@ impl PrimeGenerator {
         sieve[1] = false;
 
         for i in 2..=((limit as f64).sqrt() as usize) {
-            if sieve[i as usize] {
-                for j in (i * i..=limit as usize).step_by(i as usize) {
+            if sieve[i] {
+                for j in (i * i..=limit as usize).step_by(i) {
                     sieve[j] = false;
                 }
             }
